chore: Change order of polling
The current order of polling may not be as sound as originally thought since the future or stream may have been completed by the time the task woke up. This will change the order so if the task is woken up and the future or stream is no longer pending, it will return its value accordingly, otherwise it would timeout if the timer elapsed.

This also changes it to properly prevent the future or stream from being polled again if it times out.